+++ spec/unit/mocks/effect.ts
@@ -1,8 +1,6 @@
-import etro from '../../../src'
-
 // eslint-disable-next-line no-unused-vars
-export function mockBaseEffect (watchPublic = false) {
-  let effect = jasmine.createSpyObj('effect', [
+export function mockBaseEffect () {
+  const effect = jasmine.createSpyObj('effect', [
     'getDefaultOptions',
     'tryAttach',
     'tryDetach'
@@ -13,10 +11,6 @@ export function mockBaseEffect (watchPublic = false) {
     // Otherwise, auto-refresh will cause errors.
     effect.movie = movie
   })
-
-  // I believe `watchPublic` needs to be called before we add the properties.
-  if (watchPublic)
-    effect = etro.watchPublic(effect) as etro.effect.Base
 
   effect.type = 'effect'
   effect.enabled = true