+++ spec/unit/mocks/layer.ts
@@ -1,9 +1,8 @@
-import etro from '../../../src'
 import { mockMovie } from './movie'
 
 // eslint-disable-next-line no-unused-vars
-export function mockBaseLayer (watchPublic = false) {
-  let layer = jasmine.createSpyObj('layer', [
+export function mockBaseLayer () {
+  const layer = jasmine.createSpyObj('layer', [
     'getDefaultOptions',
     'tryAttach',
     'tryDetach',
@@ -17,10 +16,6 @@ export function mockBaseLayer (watchPublic = false) {
     // Otherwise, auto-refresh will cause errors.
     layer.movie = movie
   })
-
-  // I believe `watchPublic` needs to be called before we add the properties.
-  if (watchPublic)
-    layer = etro.watchPublic(layer) as etro.layer.Base
 
   layer.type = 'layer'
   layer.active = false