+++ spec/unit/mocks/movie.ts
@@ -1,9 +1,8 @@
-import etro from '../../../src'
 import { mockAudioContext, mockCanvas } from './dom'
 
 // eslint-disable-next-line no-unused-vars
-export function mockMovie (watchPublic = false) {
-  let movie = jasmine.createSpyObj('movie', [
+export function mockMovie () {
+  const movie = jasmine.createSpyObj('movie', [
     'getDefaultOptions',
     'addLayer',
     'addEffect'
@@ -11,10 +10,6 @@ export function mockMovie (watchPublic = false) {
   movie.getDefaultOptions.and.returnValue({
     canvas: null
   })
-
-  // I believe `watchPublic` needs to be called before we add the properties.
-  if (watchPublic)
-    movie = etro.watchPublic(movie) as etro.Movie
 
   movie.enabled = true
   movie.type = 'movie'