+++ src/effect/stack.ts
@@ -51,7 +51,6 @@ export class Stack extends Visual {
 
     this.effects = new StackEffects(options.effects, this)
     options.effects.forEach(effect => this.effects.push(effect))
-    // TODO: Propagate 'change' events from children up
   }
 
   attach (movie: Movie): void {