+++ src/object.ts
@@ -5,12 +5,6 @@ export default interface EtroObject {
   currentTime: number
   /** Used in etro internals */
   type: string
-  /**
-   * Which properties to not watch for changes, for `Movie#autoRefresh`
-   *
-   * @deprecated `Movie#autoRefresh` is deprecated
-   */
-  publicExcludes: string[]
   /** Map of property name to function to run on result of `val` */
   propertyFilters: Record<string, <T>(value: T) => T>
   /**